+++ pastel/fingerprint_test.go
@@ -50,7 +50,7 @@ func genfingerAndScoresFunc() *DDAndFingerprints {
 			Sexy:     0.5,
 		},
 
-		ImageFingerprintOfCandidateImageFile: []float32{1, 2, 3},
+		ImageFingerprintOfCandidateImageFile: []float64{1, 2, 3},
 
 		HashOfCandidateImageFile: "HashOfCandidateImageFile",
 	}
